job_portal/requirements.txt
Django==5.2.8
djangorestframework==3.15.2
django-cors-headers==4.6.0
djangorestframework-simplejwt==5.3.0
Pillow==10.4.0
psycopg2-binary==2.9.9